+++ TablePro/AppDelegate.swift
@@ -166,11 +166,6 @@ class AppDelegate: NSObject, NSApplicationDelegate {
         SQLFolderWatcher.shared.reload()
     }
 
-    @objc func showHelp(_ sender: Any?) {
-        if let url = URL(string: "https://docs.tablepro.app") {
-            NSWorkspace.shared.open(url)
-        }
-    }
 
     // MARK: - Window Notifications